+++ include/numkong/spatials/haswell.h
@@ -78,11 +78,11 @@ nk_define_cross_normalized_symmetric_(euclidean, f16, haswell, f16, f32, /*norm_
                                       nk_dots_reduce_sumsq_f16_, nk_load_b128_haswell_, nk_partial_load_b32x4_haswell_,
                                       nk_store_b128_haswell_, nk_partial_store_b32x4_haswell_, 1)
 
-nk_define_cross_normalized_packed_(angular, bf16, haswell, bf16, f32, f32, /*norm_value_type=*/f32, f32, nk_b128_vec_t,
+nk_define_cross_normalized_packed_(angular, bf16, haswell, bf16, bf16, f32, /*norm_value_type=*/f32, f32, nk_b128_vec_t,
                                    nk_dots_packed_bf16_haswell, nk_angular_through_f32_from_dot_haswell_,
                                    nk_dots_reduce_sumsq_bf16_, nk_load_b128_haswell_, nk_partial_load_b32x4_haswell_,
                                    nk_store_b128_haswell_, nk_partial_store_b32x4_haswell_, 1)
-nk_define_cross_normalized_packed_(euclidean, bf16, haswell, bf16, f32, f32, /*norm_value_type=*/f32, f32,
+nk_define_cross_normalized_packed_(euclidean, bf16, haswell, bf16, bf16, f32, /*norm_value_type=*/f32, f32,
                                    nk_b128_vec_t, nk_dots_packed_bf16_haswell,
                                    nk_euclidean_through_f32_from_dot_haswell_, nk_dots_reduce_sumsq_bf16_,
                                    nk_load_b128_haswell_, nk_partial_load_b32x4_haswell_, nk_store_b128_haswell_,
